proj/python/requirements.txt
pillow==9.4.0
pyautogui==0.9.53
pytesseract==0.3.10
opencv-python==4.7.0.72
numpy==1.24.2
pywin32==305
psutil==5.9.4